// src/domain/charity/enums/charity-action-type.enum.ts
export enum CharityActionType {
  SADAQA = 'SADAQA',
  MARAUDE = 'MARAUDE',
  COLLECTE = 'COLLECTE',
  DISTRIBUTION = 'DISTRIBUTION',
  VISITE = 'VISITE',
  PARRAINAGE = 'PARRAINAGE',
  MENTORAT = 'MENTORAT',
  RANDONNEE_SOLIDAIRE = 'RANDONNEE_SOLIDAIRE',
  COURSE_SOLIDAIRE = 'COURSE_SOLIDAIRE',
  VOYAGE_SOLIDAIRE = 'VOYAGE_SOLIDAIRE',
  SICK_ISOLATED_VISIT = 'SICK_ISOLATED_VISIT',
  FAMILY_VISIT = 'FAMILY_VISIT',
  FRATERNAL_VISIT = 'FRATERNAL_VISIT',
  EDUCATION = 'EDUCATION',
  TEACHING = 'TEACHING',
  CONFERENCE_VIGIL = 'CONFERENCE_VIGIL',
  SPIRITUAL_REMINDER = 'SPIRITUAL_REMINDER',
  ASSOCIATIVE_ENGAGEMENT = 'ASSOCIATIVE_ENGAGEMENT',
  COMMUNITY_ENGAGEMENT = 'COMMUNITY_ENGAGEMENT',
  OTHER = 'OTHER',
}